Enum itsybitsy_m0::atsamd21g18a::adc::inputctrl::MUXPOSR

pub enum MUXPOSR {
    PIN0,
    PIN1,
    PIN2,
    PIN3,
    PIN4,
    PIN5,
    PIN6,
    PIN7,
    PIN8,
    PIN9,
    PIN10,
    PIN11,
    PIN12,
    PIN13,
    PIN14,
    PIN15,
    PIN16,
    PIN17,
    PIN18,
    PIN19,
    TEMP,
    BANDGAP,
    SCALEDCOREVCC,
    SCALEDIOVCC,
    DAC,
    _Reserved(u8),
}

Possible values of the field MUXPOS

Variants
PIN0

ADC AIN0 Pin

PIN1

ADC AIN1 Pin

PIN2

ADC AIN2 Pin

PIN3

ADC AIN3 Pin

PIN4

ADC AIN4 Pin

PIN5

ADC AIN5 Pin

PIN6

ADC AIN6 Pin

PIN7

ADC AIN7 Pin

PIN8

ADC AIN8 Pin

PIN9

ADC AIN9 Pin

PIN10

ADC AIN10 Pin

PIN11

ADC AIN11 Pin

PIN12

ADC AIN12 Pin

PIN13

ADC AIN13 Pin

PIN14

ADC AIN14 Pin

PIN15

ADC AIN15 Pin

PIN16

ADC AIN16 Pin

PIN17

ADC AIN17 Pin

PIN18

ADC AIN18 Pin

PIN19

ADC AIN19 Pin

TEMP

Temperature Reference

BANDGAP

Bandgap Voltage

SCALEDCOREVCC

1/4 Scaled Core Supply

SCALEDIOVCC

1/4 Scaled I/O Supply

DAC

DAC Output

_Reserved(u8)

Reserved

Methods

impl MUXPOSR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_pin0(&self) -> bool

Checks if the value of the field is PIN0

pub fn is_pin1(&self) -> bool

Checks if the value of the field is PIN1

pub fn is_pin2(&self) -> bool

Checks if the value of the field is PIN2

pub fn is_pin3(&self) -> bool

Checks if the value of the field is PIN3

pub fn is_pin4(&self) -> bool

Checks if the value of the field is PIN4

pub fn is_pin5(&self) -> bool

Checks if the value of the field is PIN5

pub fn is_pin6(&self) -> bool

Checks if the value of the field is PIN6

pub fn is_pin7(&self) -> bool

Checks if the value of the field is PIN7

pub fn is_pin8(&self) -> bool

Checks if the value of the field is PIN8

pub fn is_pin9(&self) -> bool

Checks if the value of the field is PIN9

pub fn is_pin10(&self) -> bool

Checks if the value of the field is PIN10

pub fn is_pin11(&self) -> bool

Checks if the value of the field is PIN11

pub fn is_pin12(&self) -> bool

Checks if the value of the field is PIN12

pub fn is_pin13(&self) -> bool

Checks if the value of the field is PIN13

pub fn is_pin14(&self) -> bool

Checks if the value of the field is PIN14

pub fn is_pin15(&self) -> bool

Checks if the value of the field is PIN15

pub fn is_pin16(&self) -> bool

Checks if the value of the field is PIN16

pub fn is_pin17(&self) -> bool

Checks if the value of the field is PIN17

pub fn is_pin18(&self) -> bool

Checks if the value of the field is PIN18

pub fn is_pin19(&self) -> bool

Checks if the value of the field is PIN19

pub fn is_temp(&self) -> bool

Checks if the value of the field is TEMP

pub fn is_bandgap(&self) -> bool

Checks if the value of the field is BANDGAP

pub fn is_scaledcorevcc(&self) -> bool

Checks if the value of the field is SCALEDCOREVCC

pub fn is_scalediovcc(&self) -> bool

Checks if the value of the field is SCALEDIOVCC

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC