Module itsybitsy_m0::atsamd21g18a::sysctrl::bod33

3.3V Brown-Out Detector (BOD33) Control

Structs

CENR
Value of the field

ENABLER
Value of the field

HYSTR
Value of the field

LEVELR
Value of the field

MODER
Value of the field

R
Value read from the register

RUNSTDBYR
Value of the field

W
Value to write to the register

_ACTIONW
Proxy

_CENW
Proxy

_ENABLEW
Proxy

_HYSTW
Proxy

_LEVELW
Proxy

_MODEW
Proxy

_PSELW
Proxy

_RUNSTDBYW
Proxy

Enums

ACTIONR
Possible values of the field ACTION

ACTIONW
Values that can be written to the field ACTION

PSELR
Possible values of the field PSEL

PSELW
Values that can be written to the field PSEL