Module itsybitsy_m0::prelude

Import the prelude to gain convenient access to helper traits

Traits

_atsamd21_hal_time_U32Ext
Extension trait that adds convenience methods to the u32 type

_embedded_hal_PwmPin
A single PWM channel / pin

_embedded_hal_blocking_delay_DelayMs
Millisecond delay

_embedded_hal_blocking_delay_DelayUs
Microsecond delay

_embedded_hal_blocking_i2c_WriteRead
Blocking write + read

_embedded_hal_blocking_i2c_Read
Blocking read

_embedded_hal_blocking_i2c_Write
Blocking write

_embedded_hal_blocking_serial_Write
Write half of a serial interface (blocking variant)

_embedded_hal_blocking_spi_Transfer
Blocking transfer

_embedded_hal_blocking_spi_Write
Blocking write

_embedded_hal_digital_OutputPin
Single digital push-pull output pin

_embedded_hal_serial_Read
Read half of a serial interface

_embedded_hal_serial_Write
Write half of a serial interface

_embedded_hal_spi_FullDuplex
Full duplex (master mode)

_embedded_hal_timer_CountDown
A count down timer