Struct itsybitsy_m0::Pins

pub struct Pins {
    pub port: Port,
    pub a0: Pa2<Input<Floating>>,
    pub a1: Pb8<Input<Floating>>,
    pub a2: Pb9<Input<Floating>>,
    pub a3: Pa4<Input<Floating>>,
    pub a4: Pa5<Input<Floating>>,
    pub a5: Pb2<Input<Floating>>,
    pub d0: Pa11<Input<Floating>>,
    pub d1: Pa10<Input<Floating>>,
    pub d2: Pa14<Input<Floating>>,
    pub d3: Pa9<Input<Floating>>,
    pub d4: Pa8<Input<Floating>>,
    pub d5: Pa15<Input<Floating>>,
    pub d7: Pa21<Input<Floating>>,
    pub d9: Pa7<Input<Floating>>,
    pub d10: Pa18<Input<Floating>>,
    pub d11: Pa16<Input<Floating>>,
    pub d12: Pa19<Input<Floating>>,
    pub d13: Pa17<Input<Floating>>,
    pub mosi: Pb10<Input<Floating>>,
    pub miso: Pa12<Input<Floating>>,
    pub sck: Pb11<Input<Floating>>,
    pub flash_mosi: Pb22<Input<Floating>>,
    pub flash_miso: Pb3<Input<Floating>>,
    pub flash_sck: Pb23<Input<Floating>>,
    pub flash_cs: Pa27<Input<Floating>>,
    pub scl: Pa23<Input<Floating>>,
    pub sda: Pa22<Input<Floating>>,
    pub dotstar_ci: Pa0<Input<Floating>>,
    pub dotstar_di: Pa1<Input<Floating>>,
    pub swdio: Pa31<Input<Floating>>,
    pub swdclk: Pa30<Input<Floating>>,
    pub usb_dm: Pa24<Input<Floating>>,
    pub usb_dp: Pa25<Input<Floating>>,
}

Maps the pins to their arduino names and the numbers printed on the board.

Fields
port: Port

Opaque port reference

a0: Pa2<Input<Floating>>

Analog pin 0. Can act as a true analog output as it has a DAC (which is not currently supported by this hal) as well as input.

a1: Pb8<Input<Floating>>

Analog Pin 1

a2: Pb9<Input<Floating>>

Analog Pin 2

a3: Pa4<Input<Floating>>

Analog Pin 3

a4: Pa5<Input<Floating>>

Analog Pin 4

a5: Pb2<Input<Floating>>

Analog Pin 5

d0: Pa11<Input<Floating>>

Pin 0, rx

d1: Pa10<Input<Floating>>

Pin 1, tx

d2: Pa14<Input<Floating>>d3: Pa9<Input<Floating>>d4: Pa8<Input<Floating>>d5: Pa15<Input<Floating>>d7: Pa21<Input<Floating>>d9: Pa7<Input<Floating>>d10: Pa18<Input<Floating>>d11: Pa16<Input<Floating>>d12: Pa19<Input<Floating>>d13: Pa17<Input<Floating>>

Digital pin number 13, which is also attached to the red LED. PWM capable.

mosi: Pb10<Input<Floating>>

The SPI MOSI

miso: Pa12<Input<Floating>>

The SPI MISO

sck: Pb11<Input<Floating>>

The SPI SCK

flash_mosi: Pb22<Input<Floating>>

The MOSI pin attached to the on-board SPI flash

flash_miso: Pb3<Input<Floating>>

The MISO pin attached to the on-board SPI flash

flash_sck: Pb23<Input<Floating>>

The SCK pin attached to the on-board SPI flash

flash_cs: Pa27<Input<Floating>>

The CS pin attached to the on-board SPI flash

scl: Pa23<Input<Floating>>

The I2C clock

sda: Pa22<Input<Floating>>

The I2C data line

dotstar_ci: Pa0<Input<Floating>>

The DotStar clock

dotstar_di: Pa1<Input<Floating>>

The DotStar data line

swdio: Pa31<Input<Floating>>swdclk: Pa30<Input<Floating>>usb_dm: Pa24<Input<Floating>>

The USB D- pad

usb_dp: Pa25<Input<Floating>>

The USB D+ pad

Methods

impl Pins

pub fn new(port: PORT) -> Self